pyproject.toml
[project]
name = "obsidian-sync"
version = "0.1.0"
description = "Bidirectional sync daemon between local Obsidian vault and Google Drive"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "pyyaml>=6.0.3",
    "watchdog>=6.0.0",
]
